requirements.txt
selenium==4.15.0
requests==2.31.0
openai==1.3.0
Pillow==10.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.0
lxml==4.9.0
pandas==2.1.0
chromedriver-autoinstaller==0.6.2
